# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pulse 109 Platform API
  version: 1.0.0
  description: >-
    Contract-first API for intake, operator assistance, incident management,
    analytics and regional system adapters. All timestamps use RFC 3339 with
    an explicit UTC offset. Human confirmation remains mandatory for routing,
    priority changes, duplicate merges and generated replies.
servers:
  - url: https://pulse109.example.gov.kz
    description: Production placeholder. Replace after infrastructure approval.
...
paths:
  /v1/appeals/preflight:
    post:
      tags: [Requests, Incidents]
      summary: Suggest possible duplicates before an appeal is submitted
      operationId: preflightAppeal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreflightRequest'
      responses:
        '200':
          description: Advisory duplicate candidates; no appeal is stored or merged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResult'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/requests:
    post:
      tags: [Requests]
      summary: Create or import an appeal
      operationId: createRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/RegionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '201':
          description: Appeal created
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
        '200':
          description: Existing appeal returned for a repeated idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /v1/incidents:
    post:
      tags: [Incidents]
      summary: Create a city incident without deleting member appeals
      operationId: createIncident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/RegionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncident'
      responses:
        '201':
          description: Incident created in proposed state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /v1/incidents/{incident_id}/members:
    post:
      tags: [Incidents]
      summary: Add an appeal to an incident after human review
      operationId: addIncidentMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_id, incident_version, decision, reason_code]
              properties:
                request_id:
                  type: string
                  format: uuid
                incident_version:
                  type: integer
                  minimum: 1
                decision:
                  type: string
                  enum: [confirm, reject, remove]
                reason_code:
                  type: string
                note:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: Membership decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentMember'
  /v1/incidents/{incident_id}/confirm:
    post:
      tags: [Incidents]
      summary: Confirm or reject a proposed incident
      operationId: confirmIncident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [incident_version, decision, reason_code]
              properties:
                incident_version:
                  type: integer
                  minimum: 1
                decision:
                  type: string
                  enum: [confirm, reject]
                reason_code:
                  type: string
                note:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Incident state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /v1/alerts:
    get:
      tags: [Analytics]
      summary: List actionable anomaly and incident alerts
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - name: status
          in: query
          schema:
            type: string
            enum: [new, acknowledged, resolved, dismissed]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
  /v1/analytics/query:
    post:
      tags: [Analytics]
      summary: Execute a governed query against the semantic metric layer
      operationId: queryAnalytics
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsQuery'
      responses:
        '200':
          description: Metric result with provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResult'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/reports:
    post:
      tags: [Reports]
      summary: Create a PDF or XLSX report asynchronously
      operationId: createReport
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
  /v1/catalog/policies:
    get:
      tags: [Catalog]
      summary: Get effective versioned routing, SLA, and confidence policies
      operationId: listPolicies
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - name: effective_at
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: include_drafts
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Effective or explicitly authorized draft policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyDefinition'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the approved government identity provider.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    RegionId:
      name: X-Region-Id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[A-Z0-9_-]{2,32}$'
...
    IncidentId:
      name: incident_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  headers:
    CorrelationId:
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Version, idempotency or state conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The verified actor lacks the required region or purpose scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unprocessable:
      description: Business validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    CreateRequest:
      type: object
      additionalProperties: false
      required:
        - source_system
        - source_request_id
        - region_id
        - received_at
        - received_at_quality
        - channel
      properties:
        source_system:
          type: string
          maxLength: 64
        source_request_id:
          type: string
          maxLength: 128
        region_id:
          type: string
        received_at:
          type: [string, 'null']
          format: date-time
        received_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
        source_timezone:
          type: [string, 'null']
        channel:
          type: string
          enum: [phone, web, mobile, telegram, whatsapp, email, walk_in, import, other]
        language:
          type: string
          enum: [kk, ru, mixed, unknown]
          default: unknown
        text:
          type: string
          maxLength: 20000
        transcript_ref:
          type: string
        media_refs:
          type: array
          maxItems: 20
          items:
            type: string
        location:
          $ref: '#/components/schemas/LocationInput'
        citizen_token:
          type: string
          description: Pseudonymous token. Direct identifiers are stored outside this API domain.
        source_payload_ref:
          type: string
        consent_or_legal_basis:
          type: string
          maxLength: 128
      allOf:
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: missing
          then:
            properties:
              received_at:
                type: 'null'
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: date_only
          then:
            properties:
              received_at:
                type: 'null'
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: exact
          then:
            required: [received_at]
            properties:
              received_at:
                type: string
                format: date-time
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: source_tz_assumed
          then:
            required: [received_at, source_timezone]
            properties:
              received_at:
                type: string
                format: date-time
              source_timezone:
                type: string
                minLength: 1
...
    Appeal:
      allOf:
        - $ref: '#/components/schemas/CreateRequest'
        - type: object
          required: [request_id, created_at, version, status]
          properties:
            request_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            version:
              type: integer
              minimum: 1
            status:
              type: string
              enum: [new, triage, assigned, in_progress, resolved, closed, reopened, cancelled]
...
    PreflightRequest:
      type: object
      additionalProperties: false
      required: [region_id, service_id, topic_id, text]
      properties:
        region_id:
          type: string
          pattern: '^[A-Z0-9_-]{2,32}$'
        service_id:
          type: string
          minLength: 1
        topic_id:
          type: string
          minLength: 1
        text:
          type: string
          minLength: 1
          maxLength: 20000
        occurred_at:
          type: [string, 'null']
          format: date-time
        occurred_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
          default: missing
        latitude:
          type: [number, 'null']
          minimum: -90
          maximum: 90
        longitude:
          type: [number, 'null']
          minimum: -180
          maximum: 180
        limit:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
    PreflightResult:
      type: object
      additionalProperties: false
      required:
        [candidates, evaluated_factors, needs_human_confirmation, automatic_merge, synthetic_only]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/DuplicateCandidate'
        evaluated_factors:
          type: array
          items:
            type: string
            enum: [category, distance, time, lexical, semantic]
        needs_human_confirmation:
          const: true
        automatic_merge:
          const: false
        synthetic_only:
          type: boolean
    CreateIncident:
      type: object
      additionalProperties: false
      required: [region_id, topic_id, member_request_ids, proposal_source]
      properties:
        region_id:
          type: string
        topic_id:
          type: string
        service_id:
          type: string
        member_request_ids:
          type: array
          minItems: 2
          uniqueItems: true
          items:
            type: string
            format: uuid
        proposal_source:
          type: string
          enum: [operator, rule, model]
        geo_id:
          type: string
        window_started_at:
          type: string
          format: date-time
        window_ended_at:
          type: string
          format: date-time
        rationale:
          type: array
          items:
            type: string
    Incident:
      type: object
      required: [incident_id, state, region_id, topic_id, member_count, version]
      properties:
        incident_id:
          type: string
          format: uuid
        state:
          type: string
          enum: [proposed, confirmed, rejected, monitoring, resolved, closed]
        region_id:
          type: string
        topic_id:
          type: string
        service_id:
          type: string
        member_count:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 1
    IncidentMember:
      type: object
      required: [incident_id, request_id, decision, decided_at]
      properties:
        incident_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [confirm, reject, remove]
        decided_at:
          type: string
          format: date-time
    Alert:
      type: object
      required: [alert_id, type, region_id, severity, status, detected_at, evidence]
      properties:
        alert_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [volume_spike, incident_growth, sla_risk, data_quality, model_drift]
        region_id:
          type: string
        severity:
          type: string
          enum: [info, warning, high, critical]
        status:
          type: string
          enum: [new, acknowledged, resolved, dismissed]
        detected_at:
          type: string
          format: date-time
        evidence:
          type: object
          additionalProperties: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
    AnalyticsQuery:
      type: object
      additionalProperties: false
      required: [metric_id, time_range]
      properties:
        metric_id:
          type: string
        dimensions:
          type: array
          maxItems: 5
          items:
            type: string
        filters:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        time_range:
          type: object
          required: [from, to]
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        granularity:
          type: string
          enum: [hour, day, week, month]
        limit:
          type: integer
          minimum: 1
          maximum: 5000
          default: 1000
    AnalyticsResult:
      type: object
      required: [metric_id, metric_version, columns, rows, computed_at, data_cutoff]
      properties:
        metric_id:
          type: string
        metric_version:
          type: string
        columns:
          type: array
          items:
            type: object
            required: [name, type]
            properties:
              name:
                type: string
              type:
                type: string
        rows:
          type: array
          items:
            type: array
            items: {}
        computed_at:
          type: string
          format: date-time
        data_cutoff:
          type: string
          format: date-time
        provenance:
          type: array
          items:
            type: string
    ReportRequest:
      type: object
      additionalProperties: false
      required: [format, template_id, query]
      properties:
        format:
          type: string
          enum: [pdf, xlsx]
        template_id:
          type: string
        query:
          $ref: '#/components/schemas/AnalyticsQuery'
        locale:
          type: string
          enum: [kk-KZ, ru-KZ]
    Job:
      type: object
      required: [job_id, type, status, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
        result_ref:
          type: [string, 'null']
        error_code:
          type: [string, 'null']
...
    PolicyDefinition:
      type: object
      additionalProperties: false
      required:
        [policy_type, region_id, version, state, parameters, synthetic_only]
      properties:
        policy_type:
          type: string
          enum: [routing, sla, confidence]
        region_id:
          type: string
        version:
          type: string
        state:
          type: string
          enum: [draft, pending_approval, approved, retired]
        effective_from:
          type: [string, 'null']
          format: date-time
        effective_to:
          type: [string, 'null']
          format: date-time
        parameters:
          type: object
          additionalProperties: true
        approval_ref:
          type: [string, 'null']
        rollback_version:
          type: [string, 'null']
        synthetic_only:
          type: boolean
...